bugfix: Fixes missing email attachments in Inbox and k-9
This is a hack, dragons there!
Simple sharing the original Intent created via IntentBuilder results in huge number of apps that aren't email clients.
To solve this we resolve a Intent for sending a single message and get all the email clients that can handle that.
From that list we build a new list of intents with the original intent to share for the specific class that can handle that per email client.
After that we make a new IntentChooser with the last intent and a list of extra intents attached with the rest.