Fix changing data not always updating correctly
Until listview in Logbook activity is updated to use a cursor that is notified from changes of the data, trusting the data is impossible.
Problems would occur when user checks a entry -> edits it -> goes back to Logbook -> returns to the same entry.
The entry would be with old values instead of the last edit made.